pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cedkv-mvp"
version = "0.1.0"
description = "CED-KV MVP bootstrap for Phase 0-2 experiments"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
cedkv-mvp = "cedkv_mvp.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
